# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthStatus {
  authenticated: boolean;
  isAdmin: boolean;
  email?: string;
  displayName?: string;
}

export interface Visitor {
  name: string;
  email: string;
  visited_at: string;
  user_agent: string;
}

export interface VisitorList {
  visitors: Visitor[];
  totalVisitors: number;
  totalVisits: number;
  recentVisitors: Visitor[];
}

export type GoogleVisitorLoginParams = {
returnTo?: string;
};

export type GoogleVisitorCallbackParams = {
code?: string;
state?: string;
error?: string;
};

export type ListVisitorsParams = {
/**
 * @maxLength 100
 */
q?: string;
};