# lib/api-zod/src/generated/types/visitor.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visitor {
  name: string;
  email: string;
  visited_at: Date;
  user_agent: string;
}